Made various additions/changes.
Added to exception class to allow for displaying data or analytic errors
in the future. Also made slight changes to the data framework; a bug fix
and changes to abstract class.

+++ qt/src/datastream.cpp
@@ -381,7 +381,7 @@ DataStream& DataStream::operator>>(QPixmap &value)
    {
       if ( type == Pixmap )
       {
-         quint16 size;
+         quint32 size;
          if ( read(&size) )
          {
             size = qFromBigEndian(size);
